lab10/requirements.txt
asgiref==3.11.1
Django==6.0.3
djangorestframework==3.17.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pygame==2.6.1
sqlparse==0.5.5
